+++ src/routes/mod.rs
@@ -1,5 +1,5 @@
 mod authorize;
-mod challenge;
+mod authentication;
 mod credentials;
 mod did;
 mod endpoint;
@@ -9,7 +9,7 @@ mod well_known_did_config;
 
 pub mod error;
 pub use authorize::*;
-pub use challenge::*;
+pub use authentication::*;
 pub use credentials::*;
 pub use did::*;
 pub use endpoint::*;
